refactor: utils 패키지 이동
