+++ back-end/src/main/java/mine/is/gpu/AdminDataLoader.java
@@ -2,7 +2,7 @@ package mine.is.gpu;
 
 import mine.is.gpu.account.domain.Administrator;
 import mine.is.gpu.account.domain.repository.AdministratorRepository;
-import mine.is.gpu.utils.Encrypt;
+import mine.is.gpu.utils.encrypt.Encrypt;
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Component;
