+++ back-end/src/main/java/mine/is/gpu/utils/encrypt/Encrypt.java
@@ -1,4 +1,4 @@
-package mine.is.gpu.utils;
+package mine.is.gpu.utils.encrypt;
 
 import java.nio.charset.StandardCharsets;
 import java.security.MessageDigest;
