+++ back-end/src/main/java/mine/is/gpu/utils/pagination/Pagination.java
@@ -1,4 +1,4 @@
-package mine.is.gpu.pagination;
+package mine.is.gpu.utils.pagination;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
