+++ back-end/src/main/java/mine/is/gpu/utils/pagination/parser/HttpRequestPaginationParser.java
@@ -1,4 +1,4 @@
-package mine.is.gpu.pagination.parser;
+package mine.is.gpu.utils.pagination.parser;
 
 import java.util.Objects;
 import org.springframework.data.domain.PageRequest;
