+++ back-end/src/main/java/mine/is/gpu/utils/pagination/parser/PaginationParser.java
@@ -1,4 +1,4 @@
-package mine.is.gpu.pagination.parser;
+package mine.is.gpu.utils.pagination.parser;
 
 import org.springframework.data.domain.Pageable;
