+++ back-end/src/test/java/mine/is/gpu/utils/EncryptTest.java
@@ -3,6 +3,7 @@ package mine.is.gpu.utils;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotEquals;
 
+import mine.is.gpu.utils.encrypt.Encrypt;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
